1-list_len.c

+++ 1-list_len.c
@@ -2,13 +2,13 @@
 #include <stdlib.h>
 
 /**
- * list_len - adds a new node at the beginning of a list_t list
+ * list_len - returns the number of elements in a linked list_t list
  * @h: pointer to the list_t list
  * Return: Always 0
  */
 size_t list_len(const list_t *h)
 {
-	size_t i = 0;
+	unsigned int i = 0;
 
 	while (h)
 	{
